backend/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
alembic==1.13.1
email-validator==2.1.1
requests==2.32.3
aiohttp==3.9.5
pillow==10.4.0
numpy==1.24.3
pandas==2.0.3
tensorflow-cpu==2.13.0
scikit-learn==1.3.2
